+++ src/test/java/im/aop/loggers/messageinterpolation/DefaultToStringStrategyFactoryTests.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -45,6 +48,8 @@ class DefaultToStringStrategyFactoryTests {
     }
   }
 
+  static interface TestInterface {}
+
   @Test
   void findOrDefault_givenNull_withNullToStringStrategies() {
     final DefaultToStringStrategyFactory factory =
@@ -75,5 +80,47 @@ class DefaultToStringStrategyFactoryTests {
             new ObjectToStringStrategy(), List.of(new TestToStringStrategy()));
     assertThat(factory.findOrDefault(new TestClass()))
         .isExactlyInstanceOf(TestToStringStrategy.class);
+  }
+
+  @Test
+  void findOrDefault_givenTestInterfaceProxy_withNullToStringStrategies() {
+    TestInterface proxy =
+        (TestInterface)
+            Proxy.newProxyInstance(
+                ClassLoader.getSystemClassLoader(),
+                new Class<?>[] {TestInterface.class},
+                new InvocationHandler() {
+
+                  @Override
+                  public Object invoke(Object proxy, Method method, Object[] args)
+                      throws Throwable {
+                    return null;
+                  }
+                });
+
+    final DefaultToStringStrategyFactory factory =
+        new DefaultToStringStrategyFactory(new ObjectToStringStrategy(), null);
+    assertThat(factory.findOrDefault(proxy)).isExactlyInstanceOf(ObjectToStringStrategy.class);
+  }
+
+  @Test
+  void findOrDefault_givenTestInterfaceProxy_withEmptyToStringStrategies() {
+    TestInterface proxy =
+        (TestInterface)
+            Proxy.newProxyInstance(
+                ClassLoader.getSystemClassLoader(),
+                new Class<?>[] {TestInterface.class},
+                new InvocationHandler() {
+
+                  @Override
+                  public Object invoke(Object proxy, Method method, Object[] args)
+                      throws Throwable {
+                    return null;
+                  }
+                });
+
+    final DefaultToStringStrategyFactory factory =
+        new DefaultToStringStrategyFactory(new ObjectToStringStrategy(), List.of());
+    assertThat(factory.findOrDefault(proxy)).isExactlyInstanceOf(ObjectToStringStrategy.class);
   }
 }