+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+- Fix ReflectionToStringStrategy getting inaccessible exception when given JDK proxy instance.
+
+### Change
 - Reformat code with Google Java style.
 
 ## [1.1.0] - 2023-05-29